+++ Podcast/Service/Services/PodcastService.cs
@@ -27,6 +27,22 @@ namespace Service.Services
         }
         public async Task CreateAsync(PodcastCreateVM request)
         {
+            if (!request.UploadImage.CheckFileType("image"))
+                throw new InvalidOperationException("Only image files are allowed.");
+
+            
+            if (!request.UploadImage.CheckFileSize(1024))
+                throw new InvalidOperationException("File size should be less than 1 MB.");
+
+            string normalizedTitle = request.Title.Trim().ToLower();
+
+            var allPodcasts = await _podcastRepository.GetAllAsync();
+
+            bool exists = allPodcasts.Any(p => p.Title.Trim().ToLower() == normalizedTitle);
+
+            if (exists)
+                throw new InvalidOperationException("A podcast with the same title already exists.");
+
             string fileName = Guid.NewGuid().ToString() + "-" + request.UploadImage.FileName;
             string filePath = _env.GenerateFilePath("assets/images/podcasts", fileName);
             using (FileStream stream = new FileStream(filePath, FileMode.Create))
@@ -54,17 +70,38 @@ namespace Service.Services
 
         public async Task EditAsync(int id, PodcastEditVM request)
         {
+            string normalizedTitle = request.Title.Trim().ToLower();
+
+            var allPodcasts = await _podcastRepository.GetAllAsync();
+
+            bool exists = allPodcasts.Any(p =>
+                p.Id != id && p.Title.Trim().ToLower() == normalizedTitle);
+
+            if (exists)
+                throw new InvalidOperationException("A podcast with the same title already exists.");
+
             var podcast = await _podcastRepository.GetByIdAsync(id);
+            if (podcast == null) throw new Exception("Podcast not found.");
+
             if (request.UploadImage != null)
             {
+                if (!request.UploadImage.CheckFileType("image"))
+                    throw new InvalidOperationException("Only image files are allowed.");
+
+                if (!request.UploadImage.CheckFileSize(1024))
+                    throw new InvalidOperationException("File size should be less than 1 MB.");
+
                 string oldFilePath = _env.GenerateFilePath("assets/images/podcasts", podcast.Image);
                 oldFilePath.DeleteFile();
+
                 string fileName = Guid.NewGuid().ToString() + "-" + request.UploadImage.FileName;
                 string filePath = _env.GenerateFilePath("assets/images/podcasts", fileName);
+
                 using (FileStream stream = new FileStream(filePath, FileMode.Create))
                 {
                     await request.UploadImage.CopyToAsync(stream);
                 }
+
                 podcast.Image = fileName;
             }
             podcast.Title = request.Title;
@@ -72,6 +109,20 @@ namespace Service.Services
             podcast.PodcastCategoryId = request.PodcastCategoryId;
             podcast.TeamMemberId = request.TeamMemberId;
             await _podcastRepository.EditAsync(podcast);
+        }
+
+        public async Task<IEnumerable<PodcastAdminVM>> FilterAsync(string searchText, List<int>? categoryIds, List<int>? teamMemberIds)
+        {
+            var podcastsDb= await _podcastRepository.FilterAsync(searchText, categoryIds, teamMemberIds);
+            var podcasts = podcastsDb.Select(p => new PodcastAdminVM
+            {
+                Id = p.Id,
+                Title = p.Title,
+                Description = p.Description,
+                Image = p.Image,
+                TeamMember = p.TeamMember
+            }).ToList();
+            return podcasts;
         }
 
         public async Task<IEnumerable<PodcastAdminVM>> FilterByCategoryAsync(string searchText,string categoryName)
@@ -158,6 +209,19 @@ namespace Service.Services
             }).ToList();
             return podcasts;
         }
+        public async Task<IEnumerable<PodcastAdminVM>> GetAllByCategorySortedByFollowCountLessShowMoreAsync(int categoryId, int skip = 0, int take = 8)
+        {
+            var podcastsDb = await _podcastRepository.GetAllByCategorySortedByFollowCountLessShowMoreAsync(categoryId, skip, take);
+            var podcasts = podcastsDb.Select(p => new PodcastAdminVM
+            {
+                Id = p.Id,
+                Title = p.Title,
+                Description = p.Description,
+                Image = p.Image,
+                TeamMember = p.TeamMember
+            }).ToList();
+            return podcasts;
+        }
 
         public async Task<PodcastAdminVM> GetByIdAsync(int id)
         {
@@ -177,6 +241,8 @@ namespace Service.Services
 
         public async Task<IEnumerable<PodcastAdminVM>> SearchByTitleAsync(string searchText)
         {
+            if (string.IsNullOrWhiteSpace(searchText))
+                return new List<PodcastAdminVM>();
             var podcastsDb = await _podcastRepository.GetAllWithConditionAsync(p=>p.Title.Trim().ToLower().Contains(searchText.Trim().ToLower()));
             var podcasts=podcastsDb.Select(p=>new PodcastAdminVM
             {
@@ -188,5 +254,25 @@ namespace Service.Services
             }).ToList();
             return podcasts;
         }
+
+        
+
+        public async Task<IEnumerable<PodcastAdminVM>> GetAllByCategorySortedByOldestAsync(int categoryId, int skip, int take)
+        {
+            var podcastsDb= await _podcastRepository.GetAllByCategorySortedByOldestAsync(categoryId, skip, take);
+            var podcasts = podcastsDb.Select(p => new PodcastAdminVM
+            {
+                Id = p.Id,
+                Title = p.Title,
+                Description = p.Description,
+                Image = p.Image,
+                TeamMember = p.TeamMember
+            }).ToList();
+            return podcasts;
+        }
+
+
+
+
     }
 }
